requirements-testing.txt
# Testing Infrastructure Dependencies
# Comprehensive testing suite for Claude Optimized Deployment

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-xdist>=3.3.0  # Parallel test execution
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-randomly>=3.12.0
pytest-repeat>=0.9.1
pytest-memray>=1.4.0  # Memory profiling

# Coverage reporting
coverage[toml]>=7.2.0
coverage-badge>=1.1.0

# Property-based testing
hypothesis>=6.80.0

# Mutation testing
mutmut>=2.4.0

# Performance testing
pytest-benchmark>=4.0.0
pytest-profiling>=1.7.0

# Security testing
bandit>=1.7.5
safety>=2.3.0
semgrep>=1.30.0
pip-audit>=2.6.0

# Load testing
locust>=2.15.0
pytest-stress>=1.0.0

# Browser/E2E testing
selenium>=4.10.0
playwright>=1.35.0
requests>=2.31.0
httpx>=0.24.0

# Test data generation
factory-boy>=3.3.0
faker>=19.0.0
mimesis>=10.1.0

# Database testing
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0
pytest-mysql>=2.3.0

# Mock and stub utilities
responses>=0.23.0
httpretty>=1.1.4
freezegun>=1.2.0
time-machine>=2.10.0

# Test utilities
testcontainers>=3.7.0  # Docker containers for testing
pexpect>=4.8.0  # Interactive testing
pytest-watch>=4.2.0  # Auto-run tests on file changes
pytest-clarity>=1.0.1  # Better test failure output
pytest-sugar>=0.9.7  # Pretty test output

# Reporting and metrics
allure-pytest>=2.13.0
pytest-reportportal>=5.2.0

# Memory and performance analysis
memory-profiler>=0.60.0
psutil>=5.9.0
py-spy>=0.3.14

# Code quality
pylint>=2.17.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0

# Test environment management
tox>=4.6.0
pytest-env>=0.8.0

# Async testing utilities
asynctest>=0.13.0
aioresponses>=0.7.4
pytest-aiohttp>=1.0.4

# Network testing
pytest-httpserver>=1.0.8
pytest-socket>=0.6.0

# Logging and debugging
pytest-loguru>=0.3.0
loguru>=0.7.0

# Configuration testing
pytest-dotenv>=0.5.2
python-dotenv>=1.0.0

# Image and file testing
pillow>=10.0.0
pytest-datafiles>=3.0.0

# Jupyter notebook testing
nbval>=0.10.0
pytest-notebook>=0.8.0

# Documentation testing
pytest-doctestplus>=1.0.0

# API testing
tavern>=2.3.0
dredd-hooks>=0.2.0

# CLI testing
click-testing>=0.1.0
typer-testing>=0.1.0

# Database migrations testing
alembic>=1.11.0
sqlalchemy-utils>=0.41.0

# Message queue testing
celery>=5.3.0
pytest-celery>=0.1.0

# Cache testing
fakeredis>=2.15.0

# File system testing
pyfakefs>=5.2.0

# Time/date testing
freezegun>=1.2.0
pytest-freezegun>=0.4.2

# Error tracking and monitoring
sentry-sdk>=1.28.0

# Development utilities
ipdb>=0.13.13
pudb>=2022.1.3